Core/DataStores: Updated to 11.0.0
Port From (https://github.com/TrinityCore/TrinityCore/commit/df511503815b7798c83acbffbb1ce5208f189ce7)

+++ Source/Game/DataStorage/Structs/A_Records.cs
@@ -85,7 +85,7 @@ namespace Game.DataStorage
         public uint Id;
         public ushort Fallback;
         public byte BehaviorTier;
-        public int BehaviorID;
+        public short BehaviorID;
         public int[] Flags = new int[2];
     }
